+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## master
 
 <!-- Your comment below here -->
+* Add GitLab Merge Request support with a new `mr` command - [@jtslear](https://github.com/jtslear) [#1516](https://github.com/danger/danger/pull/1516)
 * Support v4 of colored2 gem to make danger work with --enable-frozen-string-literal RUBYOPT of newer Ruby versions - [@eiskrenkov](https://github.com/eiskrenkov) [#1513](https://github.com/danger/danger/pull/1513)
 * Support ruby-git 2.x - [@manicmaniac](https://github.com/manicmaniac) [#1511](https://github.com/danger/danger/pull/1511)
 * Allow terminal-table versions through 4.x - [@murilooon](https://github.com/murilooon)